refactor: folder structure for authentication components
