+++ app/routes/welcome.tsx
@@ -1,6 +1,6 @@
 import type { MetaFunction } from "remix";
 
-import { Welcome } from "front/welcome/Welcome";
+import { Welcome } from "front/authentication/Welcome";
 
 export const meta: MetaFunction = () => ({
   title: "TLM | Welcome",
